app/requirements.txt
image 
matplotlib
#torch
#torchvision
numpy 
#opencv-python
moviepy
pydub
pytube3
cloudinary
Flask
flask-restful

